+++ src/components/login-page/LoginSection/LoginSection.module.scss
@@ -6,37 +6,27 @@
   align-items: center;
   justify-content: center;
   min-height: 100vh;
- 
   padding: 10px;
-  @media (min-height: 1024px) {
-    padding: 50px 16px;
-  }
 
-  .form{
+  .wrapper{
     border-radius: 8px;
     background: #d2d2d2;
     color: #020202;
     display: flex;
     flex-direction: column;
     align-items: center;
-    gap: 18px;
-    padding: 20px 50px;
-
+    padding: 30px 50px;
+    gap: 36px;
     @media (min-height: 800px) {
       padding: 50px;
       gap: 36px;
     }
-
     .title{
       text-align: center;
       & > h2{
         @include mix.apply-font-styles(var.$h1-font);
         font-weight: 600;
-        font-size: 34px;
-
-        @media (min-height: 1024px) {
-          font-size: 40px;
-        }
+        font-size: 40px;
       }
   
       & > p{
@@ -44,9 +34,8 @@
         font-weight: 400;
         font-size: 20px;
       }
-
-
     }
+
     .link{
       @include mix.apply-font-styles(var.$h2-font);
       font-weight: 400;
@@ -64,90 +53,6 @@
     }
     &  a:hover{
       opacity: .7;
-    }
-
-    .list {
-      display: flex;
-      flex-direction: column;
-      //gap: 20px;
-      gap: 14px;
-
-      @media (min-height: 1024px) {
-        // height: clamp(550px, 86vh, 860px);
-        gap: 20px;
-      }
-
-      .list_item{
-        position: relative;
-        .btn{
-          position: absolute;
-          right: 10px;
-          top: 40px;
-        }
-      }
-
-      .item {
-        width: clamp(280px, 84vw, 360px);
-        display: flex;
-        flex-direction: column;
-        position: relative;
-        & > input{
-          padding: 10px 35px 10px 10px;
-        }
-        & > label{
-          font-weight: 600;
-          font-size: 18px;
-        }
-
-        .btn_option{
-          display: flex;
-          align-items: center;
-          cursor: pointer;
-          gap: 12px;
-          @include mix.apply-font-styles(var.$h3-font);
-          font-weight: 600;
-          font-size: 18px;
-  
-          &.agree{
-            @include mix.apply-font-styles(var.$text-16-font);
-            font-size: 16px;
-          }
-  
-          & input{
-            display: none;
-          }
-  
-          .check{
-            width: 24px;
-            height: 24px;
-            border-radius: 2px;
-            border: 1px solid var.$secondary-gray;
-           
-            & svg{
-              opacity: 0;
-              display: block;
-              border-radius: 2px;
-              width: 24px;
-              height: 24px;
-            }
-  
-            &._active{
-              border: none;
-  
-              & svg{
-                opacity: 1;
-              }
-            }
-          }
-        }
-
-      }
-
-      .error_modal{
-        @include mix.apply-font-styles(var.$text-16-font);
-        font-size: 14px;
-        color: var.$secondary-error;
-      }
     }
   }
 }
