+++ src/components/shared/UseAlert/UseAlert.jsx
@@ -7,7 +7,7 @@ import stateUseAlert from '@/src/state/stateUseAlert';
 import { useEffect } from 'react';
 import { createPortal } from "react-dom";
 
-export default function UseAlert(){
+export default function UseAlert({title='',text=''}){
   const t = useTranslations("Alert");
   const isOpen = stateUseAlert(state => state.isOpen);
   const type = stateUseAlert(state => state.type);
@@ -32,8 +32,8 @@ export default function UseAlert(){
   return createPortal(
     <div className={styles.wrapper}>
     <div className={clsx(styles.alert, styles[type])}>
-        <h2>{t(`title_${type}`)}</h2> 
-        { isText(type) && <p>{t(`text_${type}`)}</p> }
+        <h2>{title ? title : t(`title_${type}`)}</h2> 
+        { isText(type) && <p>{text ? text : t(`text_${type}`)}</p> }
 
       <CloseBtn className={styles.close_btn} onClick={close}/>
     </div>
